+++ dmr/plugins/pydantic/serializer.py
@@ -7,6 +7,7 @@ from typing import (
     ClassVar,
     Literal,
     TypeAlias,
+    TypeVar,
     Union,
     final,
 )
@@ -40,10 +41,10 @@ _IncEx: TypeAlias = (
 
 
 @final
-class ModelDumpKwargs(TypedDict, total=False):
+class ToJsonKwargs(TypedDict, total=False):
     """Keyword arguments for pydantic's model dump method."""
 
-    mode: Literal['json', 'python'] | str
+    # `mode` is expicitly left out. It is always `json`.
     include: _IncEx | None
     exclude: _IncEx | None
     context: Any | None
@@ -59,11 +60,11 @@ class ModelDumpKwargs(TypedDict, total=False):
 
 
 @final
-class FromPythonKwargs(TypedDict, total=False):
+class ToModelKwargs(TypedDict, total=False):
     """Keyword arguments for pydantic's python object validation method."""
 
+    # `from_attributes` is expicitly left out. It is always `False`.
     extra: ExtraValues | None
-    from_attributes: bool | None
     context: Any | None
     experimental_allow_partial: bool | Literal['off', 'on', 'trailing-strings']
     by_alias: bool | None
@@ -81,6 +82,8 @@ class PydanticEndpointOptimizer(BaseEndpointOptimizer):
         # during import time and cache them for later use in runtime.
         for response in metadata.responses.values():
             _get_cached_type_adapter(response.return_type)
+        # It is used in many places:
+        _get_cached_type_adapter(Any)
 
 
 class PydanticSerializer(BaseSerializer):
@@ -94,6 +97,12 @@ class PydanticSerializer(BaseSerializer):
 
         pip install 'django-modern-rest[pydantic]'
 
+    Attributes:
+        to_json_kwargs: Dictionary of kwargs that will be passed
+            to model serialization callbacks.
+        to_model_kwargs: Dictionary of kwargs that will be passed
+            to model deserialization callbacks.
+
     """
 
     __slots__ = ()
@@ -104,12 +113,11 @@ class PydanticSerializer(BaseSerializer):
     schema_generator = PydanticSchemaGenerator
 
     # Custom API:
-
-    model_dump_kwargs: ClassVar[ModelDumpKwargs] = {
+    to_json_kwargs: ClassVar[ToJsonKwargs] = {
         'by_alias': True,
-        'mode': 'json',
     }
-    from_python_kwargs: ClassVar[FromPythonKwargs] = {
+
+    to_model_kwargs: ClassVar[ToModelKwargs] = {
         'by_alias': True,
     }
 
@@ -135,13 +143,20 @@ class PydanticSerializer(BaseSerializer):
     def serialize_hook(cls, to_serialize: Any) -> Any:
         """Customize how some objects are serialized into simple objects."""
         if isinstance(to_serialize, pydantic.BaseModel):
-            return to_serialize.model_dump(**cls.model_dump_kwargs)
+            return to_serialize.model_dump(mode='json', **cls.to_json_kwargs)
+        # We support dataclasses here, because raw `JsonRenderer`
+        # does not support them, however, we use them in multiple places inside:
         if is_dataclass(to_serialize):
             return _get_cached_type_adapter(
                 type(to_serialize),  # type: ignore[arg-type]
             ).dump_python(
                 to_serialize,
             )
+        # This is a pydantic field inside a `TypedDict`, `@dataclass`, etc:
+        if hasattr(to_serialize, '__get_pydantic_core_schema__'):
+            return _get_cached_type_adapter(
+                type(to_serialize),  # type: ignore[arg-type]
+            ).dump_python(to_serialize, mode='json', **cls.to_json_kwargs)
         return super().serialize_hook(to_serialize)
 
     @override
@@ -203,7 +218,7 @@ class PydanticSerializer(BaseSerializer):
         return adapter.validate_python(
             unstructured,
             strict=strict,
-            **cls.from_python_kwargs,
+            **cls.to_model_kwargs,
         )
 
     @override
@@ -224,8 +239,7 @@ class PydanticSerializer(BaseSerializer):
         """
         return _get_cached_type_adapter(Any).dump_python(
             structured,
-            # To be in sync with `msgspec`:
-            mode='json',
+            **cls.to_json_kwargs,
         )
 
     @override
@@ -281,7 +295,11 @@ class PydanticFastSerializer(PydanticSerializer):
         *renderer* parameter is always ignored.
         """
         try:
-            return _get_cached_type_adapter(Any).dump_json(structure)
+            return _get_cached_type_adapter(Any).dump_json(
+                structure,
+                fallback=cls.serialize_hook,
+                **cls.to_json_kwargs,  # type: ignore[misc]
+            )
         except pydantic_core.PydanticSerializationError as exc:
             raise DataRenderingError(str(exc)) from exc
 
@@ -301,7 +319,10 @@ class PydanticFastSerializer(PydanticSerializer):
         *parser* parameter is always ignored.
         """
         try:
-            return _get_cached_type_adapter(Any).validate_json(buffer)
+            return _get_cached_type_adapter(Any).validate_json(
+                buffer,
+                **cls.to_model_kwargs,
+            )
         except pydantic_core.ValidationError as exc:
             raise DataParsingError(exc.errors()[0]['msg']) from exc
 
@@ -316,8 +337,11 @@ class PydanticFastSerializer(PydanticSerializer):
         return pluggable.content_type == 'application/json'
 
 
+_ModelT = TypeVar('_ModelT')
+
+
 @lru_cache(maxsize=MAX_CACHE_SIZE)
-def _get_cached_type_adapter(model: Any) -> pydantic.TypeAdapter[Any]:
+def _get_cached_type_adapter(model: _ModelT) -> pydantic.TypeAdapter[_ModelT]:
     """
     It is expensive to create, reuse existing ones.