+++ tests/test_unit/test_plugins/test_pydantic/test_complex_pydantic_fields.py
@@ -10,13 +10,12 @@ from django.conf import LazySettings
 from django.http import HttpResponse
 from faker import Faker
 from inline_snapshot import snapshot
+from pydantic_extra_types import Color
+from typing_extensions import TypedDict
 
 from dmr import Body, Controller
-from dmr.plugins.pydantic import (
-    ModelDumpKwargs,
-    PydanticFastSerializer,
-    PydanticSerializer,
-)
+from dmr.plugins.pydantic import PydanticFastSerializer, PydanticSerializer
+from dmr.plugins.pydantic.serializer import ToJsonKwargs
 from dmr.test import DMRRequestFactory
 
 
@@ -27,6 +26,7 @@ class _BodyModel(pydantic.BaseModel):
     created_at: dt.datetime
     elapsed: dt.timedelta
     url: pydantic.HttpUrl
+    color: Color
     extra: pydantic.Json[Any]
 
 
@@ -54,6 +54,7 @@ def test_complex_pydantic_serialization(
         'created_at': faker.future_datetime(),
         'elapsed': faker.time_delta(),
         'url': faker.url(),
+        'color': 'red',
         'extra': '{"key": "value"}',
     }
 
@@ -62,38 +63,42 @@ def test_complex_pydantic_serialization(
     response = _ComplexFieldsController.as_view()(request)
 
     assert isinstance(response, HttpResponse)
-    assert response.status_code == HTTPStatus.CREATED
+    assert response.status_code == HTTPStatus.CREATED, response.content
     assert json.loads(response.content).keys() == request_data.keys()
     assert json.loads(response.content)['extra'] == {'key': 'value'}
 
 
-@final
-class _RoundTripPydanticSerializer(PydanticSerializer):
-    model_dump_kwargs: ClassVar[ModelDumpKwargs] = {
-        **PydanticSerializer.model_dump_kwargs,
-        'round_trip': True,
-    }
-
-
-@final
-class _RoundTripJsonFieldController(
-    Controller[_RoundTripPydanticSerializer],
-):
-    def post(self, parsed_body: Body[_BodyModel]) -> _BodyModel:
-        return parsed_body
-
-
+@pytest.mark.parametrize(
+    'serializer',
+    [PydanticSerializer, PydanticFastSerializer],
+)
 def test_pydantic_round_trip_json_field(
     dmr_rf: DMRRequestFactory,
     faker: Faker,
+    *,
+    serializer: type[PydanticSerializer],
 ) -> None:
     """Ensures by round trip json field works."""
+
+    class _RoundTripPydanticSerializer(serializer):  # type: ignore[valid-type, misc]
+        to_json_kwargs: ClassVar[ToJsonKwargs] = {
+            **serializer.to_json_kwargs,
+            'round_trip': True,
+        }
+
+    class _RoundTripJsonFieldController(
+        Controller[_RoundTripPydanticSerializer],
+    ):
+        def post(self, parsed_body: Body[_BodyModel]) -> _BodyModel:
+            return parsed_body
+
     request_data = {
         'uid': uuid.uuid4(),
         'email': faker.email(),
         'created_at': faker.future_datetime(),
         'elapsed': faker.time_delta(),
         'url': faker.url(),
+        'color': '#FFFFFF',
         'extra': '{"key": "value"}',
     }
 
@@ -102,9 +107,56 @@ def test_pydantic_round_trip_json_field(
     response = _RoundTripJsonFieldController.as_view()(request)
 
     assert isinstance(response, HttpResponse)
-    assert response.status_code == HTTPStatus.CREATED
+    assert response.status_code == HTTPStatus.CREATED, response.content
     assert json.loads(response.content).keys() == request_data.keys()
     assert json.loads(response.content)['extra'] == '{"key":"value"}'
+
+
+@final
+class _BodyDict(TypedDict):
+    uid: uuid.UUID
+    email: pydantic.EmailStr
+    url: pydantic.AnyHttpUrl
+    token: pydantic.SecretStr
+    color: Color
+
+
+@pytest.mark.parametrize(
+    'serializer',
+    [PydanticSerializer, PydanticFastSerializer],
+)
+def test_complex_pydantic_dict(
+    dmr_rf: DMRRequestFactory,
+    faker: Faker,
+    *,
+    serializer: type[PydanticSerializer],
+) -> None:
+    """Ensures by default all complex fields work."""
+
+    class _ComplexFieldsController(
+        Controller[serializer],  # type: ignore[valid-type]
+    ):
+        def post(self, parsed_body: Body[_BodyDict]) -> _BodyDict:
+            return parsed_body
+
+    request_data = {
+        'uid': str(uuid.uuid4()),
+        'email': faker.email(),
+        'url': faker.url(),
+        'token': faker.text(),
+        'color': 'red',
+    }
+
+    request = dmr_rf.post('/whatever/', data=request_data)
+
+    response = _ComplexFieldsController.as_view()(request)
+
+    assert isinstance(response, HttpResponse)
+    assert response.status_code == HTTPStatus.CREATED, response.content
+    assert json.loads(response.content) == {
+        **request_data,
+        'token': '**********',
+    }
 
 
 @final
@@ -120,9 +172,7 @@ class _ArbitraryTypesModel(pydantic.BaseModel):
 
 
 @final
-class _ArbitraryTypesInputController(
-    Controller[PydanticSerializer],
-):
+class _ArbitraryTypesInputController(Controller[PydanticSerializer]):
     def post(self, parsed_body: Body[_ArbitraryTypesModel]) -> int:
         raise NotImplementedError