+++ tests/test_unit/test_plugins/test_pydantic/test_pydantic_fast.py
@@ -10,9 +10,7 @@ from inline_snapshot import snapshot
 
 from dmr import Body, Controller
 from dmr.exceptions import EndpointMetadataError
-from dmr.plugins.pydantic import (
-    PydanticFastSerializer,
-)
+from dmr.plugins.pydantic import PydanticFastSerializer
 from dmr.test import DMRRequestFactory
 from tests.infra.xml_format import XmlParser, XmlRenderer