+++ tests/test_unit/test_plugins/test_pydantic/test_pydantic_snapshots.py
@@ -2,6 +2,8 @@ import json
 
 import pydantic
 from django.urls import path
+from pydantic_extra_types import Color
+from pydantic_extra_types.phone_numbers import PhoneNumber
 from syrupy.assertion import SnapshotAssertion
 
 from dmr import Body, Controller
@@ -13,6 +15,8 @@ from dmr.routing import Router
 class _UserModel(pydantic.BaseModel):
     email: pydantic.EmailStr
     password: pydantic.SecretStr
+    phone: PhoneNumber
+    preffered_color: Color
 
 
 class _UserController(Controller[PydanticSerializer]):